+++ docs/_posts/2024-xx-xx-v4.11.0.md
@@ -122,6 +122,7 @@ It is also available through [Artifact Hub](https://artifacthub.io/packages/helm
 * Fix severity not being set for vulnerabilities from VulnDB - [apiserver/#3595]
 * Fix `JDOFatalUserException` for long reference URLs from OSS Index - [apiserver/#3650]
 * Fix unhandled `ClientErrorException`s causing a `HTTP 500` response - [apiserver/#3659]
+* Fix unique constraint violation during NVD mirroring via feed files - [apiserver/#3664]
 * Fix `VUE_APP_SERVER_URL` being ignored - [frontend/#682]
 * Fix visibility of "Vulnerabilities" and "Policy Violations" columns not being toggle-able individually - [frontend/#686]
 * Fix finding search routes - [frontend/#689]
@@ -252,6 +253,7 @@ Special thanks to everyone who contributed code to implement enhancements and fi
 [apiserver/#3657]: https://github.com/DependencyTrack/dependency-track/pull/3657
 [apiserver/#3659]: https://github.com/DependencyTrack/dependency-track/pull/3659
 [apiserver/#3661]: https://github.com/DependencyTrack/dependency-track/pull/3661
+[apiserver/#3664]: https://github.com/DependencyTrack/dependency-track/pull/3664
 
 [frontend/#682]: https://github.com/DependencyTrack/frontend/pull/682
 [frontend/#683]: https://github.com/DependencyTrack/frontend/pull/683
