+++ src/main/java/org/dependencytrack/tasks/NistApiMirrorTask.java
@@ -33,7 +33,6 @@ import io.github.jeremylong.openvulnerability.client.nvd.NvdCveClient;
 import io.github.jeremylong.openvulnerability.client.nvd.NvdCveClientBuilder;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.concurrent.BasicThreadFactory;
-import org.apache.commons.lang3.tuple.Pair;
 import org.apache.hc.client5.http.auth.AuthScope;
 import org.apache.hc.client5.http.auth.NTCredentials;
 import org.apache.hc.client5.http.auth.UsernamePasswordCredentials;
@@ -45,27 +44,20 @@ import org.apache.hc.client5.http.impl.routing.SystemDefaultRoutePlanner;
 import org.dependencytrack.common.AlpineHttpProxySelector;
 import org.dependencytrack.event.EpssMirrorEvent;
 import org.dependencytrack.event.IndexEvent;
-import org.dependencytrack.event.IndexEvent.Action;
 import org.dependencytrack.event.NistApiMirrorEvent;
 import org.dependencytrack.model.AffectedVersionAttribution;
 import org.dependencytrack.model.Vulnerability;
-import org.dependencytrack.model.Vulnerability.Source;
 import org.dependencytrack.model.VulnerableSoftware;
 import org.dependencytrack.persistence.QueryManager;
+import org.dependencytrack.persistence.listener.IndexingInstanceLifecycleListener;
+import org.dependencytrack.persistence.listener.L2CacheEvictingInstanceLifecycleListener;
 import org.dependencytrack.util.DebugDataEncryption;
-import org.dependencytrack.util.PersistenceUtil.Diff;
-import org.dependencytrack.util.PersistenceUtil.Differ;
 
-import javax.jdo.Query;
 import java.time.Duration;
 import java.time.Instant;
 import java.time.ZoneOffset;
 import java.time.ZonedDateTime;
-import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.ThreadPoolExecutor;
@@ -73,22 +65,21 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 
 import static io.github.jeremylong.openvulnerability.client.nvd.NvdCveClientBuilder.aNvdCveApi;
-import static java.util.stream.Collectors.groupingBy;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
 import static org.datanucleus.PropertyNames.PROPERTY_PERSISTENCE_BY_REACHABILITY_AT_COMMIT;
+import static org.datanucleus.PropertyNames.PROPERTY_RETAIN_VALUES;
 import static org.dependencytrack.model.ConfigPropertyConstants.VULNERABILITY_SOURCE_NVD_API_KEY;
 import static org.dependencytrack.model.ConfigPropertyConstants.VULNERABILITY_SOURCE_NVD_API_LAST_MODIFIED_EPOCH_SECONDS;
 import static org.dependencytrack.model.ConfigPropertyConstants.VULNERABILITY_SOURCE_NVD_API_URL;
 import static org.dependencytrack.parser.nvd.api20.ModelConverter.convert;
 import static org.dependencytrack.parser.nvd.api20.ModelConverter.convertConfigurations;
-import static org.dependencytrack.util.PersistenceUtil.assertPersistent;
 
 /**
  * A {@link Subscriber} that mirrors the content of the NVD through the NVD API 2.0.
  *
  * @since 4.10.0
  */
-public class NistApiMirrorTask implements Subscriber {
+public class NistApiMirrorTask extends AbstractNistMirrorTask implements Subscriber {
 
     private static final Logger LOGGER = Logger.getLogger(NistApiMirrorTask.class);
 
@@ -175,12 +166,15 @@ public class NistApiMirrorTask implements Subscriber {
                         executor.submit(() -> {
                             try (final var qm = new QueryManager().withL2CacheDisabled()) {
                                 qm.getPersistenceManager().setProperty(PROPERTY_PERSISTENCE_BY_REACHABILITY_AT_COMMIT, "false");
+                                qm.getPersistenceManager().setProperty(PROPERTY_RETAIN_VALUES, "true");
+                                qm.getPersistenceManager().addInstanceLifecycleListener(new IndexingInstanceLifecycleListener(Event::dispatch),
+                                        Vulnerability.class, VulnerableSoftware.class);
+                                qm.getPersistenceManager().addInstanceLifecycleListener(new L2CacheEvictingInstanceLifecycleListener(qm),
+                                        AffectedVersionAttribution.class, Vulnerability.class, VulnerableSoftware.class);
 
-                                // Note: persistentVuln is in HOLLOW state (all fields except ID are unloaded).
-                                // https://www.datanucleus.org/products/accessplatform_6_0/jdo/persistence.html#lifecycle
                                 final Vulnerability persistentVuln = synchronizeVulnerability(qm, vuln);
                                 synchronizeVulnerableSoftware(qm, persistentVuln, vsList);
-                            } catch (Exception ex) {
+                            } catch (RuntimeException ex) {
                                 LOGGER.error("An unexpected error occurred while processing %s".formatted(vuln.getVulnId()), ex);
                             } finally {
                                 final int currentNumMirrored = numMirrored.incrementAndGet();
@@ -225,147 +219,11 @@ public class NistApiMirrorTask implements Subscriber {
         }
 
         if (updateLastModified(lastModified)) {
-            Event.dispatch(new IndexEvent(Action.COMMIT, Vulnerability.class));
+            Event.dispatch(new IndexEvent(IndexEvent.Action.COMMIT, Vulnerability.class));
+            Event.dispatch(new IndexEvent(IndexEvent.Action.COMMIT, VulnerableSoftware.class));
         }
 
         Event.dispatch(new EpssMirrorEvent());
-    }
-
-    private static Vulnerability synchronizeVulnerability(final QueryManager qm, final Vulnerability vuln) {
-        final Pair<Vulnerability, IndexEvent> vulnIndexEventPair = qm.runInTransaction(trx -> {
-            trx.setSerializeRead(true); // SELECT ... FOR UPDATE
-
-            Vulnerability persistentVuln = getVulnerabilityByCveId(qm, vuln.getVulnId());
-            if (persistentVuln == null) {
-                persistentVuln = qm.getPersistenceManager().makePersistent(vuln);
-                return Pair.of(persistentVuln, new IndexEvent(Action.CREATE, persistentVuln));
-            } else {
-                final Map<String, Diff> diffs = updateVulnerability(persistentVuln, vuln);
-                if (!diffs.isEmpty()) {
-                    LOGGER.debug("%s has changed: %s".formatted(vuln.getVulnId(), diffs));
-                    return Pair.of(persistentVuln, new IndexEvent(Action.UPDATE, persistentVuln));
-                }
-
-                LOGGER.debug("%s has not changed".formatted(vuln.getVulnId()));
-                return Pair.of(persistentVuln, null);
-            }
-        });
-
-        final IndexEvent indexEvent = vulnIndexEventPair.getRight();
-        final Vulnerability persistentVuln = vulnIndexEventPair.getLeft();
-
-        if (indexEvent != null) {
-            Event.dispatch(indexEvent);
-        }
-
-        return persistentVuln;
-    }
-
-    private static void synchronizeVulnerableSoftware(final QueryManager qm, final Vulnerability persistentVuln, final List<VulnerableSoftware> vsList) {
-        qm.runInTransaction(tx -> {
-            tx.setSerializeRead(false);
-
-            // Get all VulnerableSoftware records that are currently associated with the vulnerability.
-            // Note: For SOME ODD REASON, duplicate (as in, same database ID and all) VulnerableSoftware
-            // records are returned, when operating on data that was originally created by the feed-based
-            // NistMirrorTask. We thus have to deduplicate here.
-            final List<VulnerableSoftware> vsOldList = persistentVuln.getVulnerableSoftware().stream().distinct().toList();
-            LOGGER.trace("%s: Existing VS: %d".formatted(persistentVuln.getVulnId(), vsOldList.size()));
-
-            // Get attributions for all existing VulnerableSoftware records.
-            final Map<Long, List<AffectedVersionAttribution>> attributionsByVsId =
-                    qm.getAffectedVersionAttributions(persistentVuln, vsOldList).stream()
-                            .collect(groupingBy(attribution -> attribution.getVulnerableSoftware().getId()));
-            for (final VulnerableSoftware vsOld : vsOldList) {
-                vsOld.setAffectedVersionAttributions(attributionsByVsId.get(vsOld.getId()));
-            }
-
-            // Based on the lists of currently reported, and previously reported VulnerableSoftware records,
-            // divide the previously reported ones into lists of records to keep, and records to remove.
-            // Records to keep are removed from vsList. Remaining records in vsList thus are entirely new.
-            final var vsListToRemove = new ArrayList<VulnerableSoftware>();
-            final var vsListToKeep = new ArrayList<VulnerableSoftware>();
-            for (final VulnerableSoftware vsOld : vsOldList) {
-                if (vsList.removeIf(vsOld::equalsIgnoringDatastoreIdentity)) {
-                    vsListToKeep.add(vsOld);
-                } else {
-                    final List<AffectedVersionAttribution> attributions = vsOld.getAffectedVersionAttributions();
-                    if (attributions == null || attributions.isEmpty()) {
-                        // DT versions prior to 4.7.0 did not record attributions.
-                        // Drop the VulnerableSoftware for now. If it was previously
-                        // reported by another source, it will be recorded and attributed
-                        // whenever that source is mirrored again.
-                        vsListToRemove.add(vsOld);
-                        continue;
-                    }
-
-                    final boolean previouslyReportedByNvd = attributions.stream()
-                            .anyMatch(attr -> attr.getSource() == Source.NVD);
-                    final boolean previouslyReportedByOthers = !previouslyReportedByNvd;
-
-                    if (previouslyReportedByOthers) {
-                        vsListToKeep.add(vsOld);
-                    } else {
-                        vsListToRemove.add(vsOld);
-                    }
-                }
-            }
-            LOGGER.trace("%s: vsListToKeep: %d".formatted(persistentVuln.getVulnId(), vsListToKeep.size()));
-            LOGGER.trace("%s: vsListToRemove: %d".formatted(persistentVuln.getVulnId(), vsListToRemove.size()));
-
-            // Remove attributions for VulnerableSoftware records that are no longer reported.
-            if (!vsListToRemove.isEmpty()) {
-                qm.deleteAffectedVersionAttributions(persistentVuln, vsListToRemove, Source.NVD);
-            }
-
-            final var attributionDate = new Date();
-
-            // For VulnerableSoftware records that existed before, update the lastSeen timestamp.
-            for (final VulnerableSoftware oldVs : vsListToKeep) {
-                oldVs.getAffectedVersionAttributions().stream()
-                        .filter(attribution -> attribution.getSource() == Source.NVD)
-                        .findAny()
-                        .ifPresent(attribution -> attribution.setLastSeen(attributionDate));
-            }
-
-            // For VulnerableSoftware records that are newly reported for this vulnerability, check if any matching
-            // records exist in the database that are currently associated with other (or no) vulnerabilities.
-            for (final VulnerableSoftware vs : vsList) {
-                final VulnerableSoftware existingVs = qm.getVulnerableSoftwareByCpe23(
-                        vs.getCpe23(),
-                        vs.getVersionEndExcluding(),
-                        vs.getVersionEndIncluding(),
-                        vs.getVersionStartExcluding(),
-                        vs.getVersionStartIncluding()
-                );
-                if (existingVs != null) {
-                    final boolean hasAttribution = qm.hasAffectedVersionAttribution(persistentVuln, existingVs, Source.NVD);
-                    if (!hasAttribution) {
-                        LOGGER.trace("%s: Adding attribution".formatted(persistentVuln.getVulnId()));
-                        final AffectedVersionAttribution attribution = createAttribution(persistentVuln, existingVs, attributionDate);
-                        qm.getPersistenceManager().makePersistent(attribution);
-                    } else {
-                        LOGGER.debug("%s: Encountered dangling attribution; Re-using by updating firstSeen and lastSeen timestamps".formatted(persistentVuln.getVulnId()));
-                        final AffectedVersionAttribution existingAttribution = qm.getAffectedVersionAttribution(persistentVuln, existingVs, Source.NVD);
-                        existingAttribution.setFirstSeen(attributionDate);
-                        existingAttribution.setLastSeen(attributionDate);
-                    }
-                    vsListToKeep.add(existingVs);
-                } else {
-                    LOGGER.trace("%s: Creating new VS".formatted(persistentVuln.getVulnId()));
-                    final VulnerableSoftware persistentVs = qm.getPersistenceManager().makePersistent(vs);
-                    final AffectedVersionAttribution attribution = createAttribution(persistentVuln, persistentVs, attributionDate);
-                    qm.getPersistenceManager().makePersistent(attribution);
-                    vsListToKeep.add(persistentVs);
-                }
-            }
-
-            LOGGER.trace("%s: Final vsList: %d".formatted(persistentVuln.getVulnId(), vsListToKeep.size()));
-            if (!Objects.equals(persistentVuln.getVulnerableSoftware(), vsListToKeep)) {
-                LOGGER.trace("%s: vsList has changed: %s".formatted(persistentVuln.getVulnId(), new Diff(persistentVuln.getVulnerableSoftware(), vsListToKeep)));
-                persistentVuln.setVulnerableSoftware(vsListToKeep);
-            }
-        });
     }
 
     private static NvdCveClient createApiClient(final String apiUrl, final String apiKey, final long lastModifiedEpochSeconds) {
@@ -407,92 +265,6 @@ public class NistApiMirrorTask implements Subscriber {
         }
 
         return true;
-    }
-
-    private static AffectedVersionAttribution createAttribution(final Vulnerability vuln, final VulnerableSoftware vs,
-                                                                final Date attributionDate) {
-        final var attribution = new AffectedVersionAttribution();
-        attribution.setSource(Source.NVD);
-        attribution.setVulnerability(vuln);
-        attribution.setVulnerableSoftware(vs);
-        attribution.setFirstSeen(attributionDate);
-        attribution.setLastSeen(attributionDate);
-        return attribution;
-    }
-
-    /**
-     * Get a {@link Vulnerability} by its CVE ID (implying the source {@link Source#NVD}).
-     * <p>
-     * It differs from {@link QueryManager#getVulnerabilityByVulnId(String, String)} in that it does not fetch any
-     * adjacent relationships (e.g. affected components and aliases).
-     *
-     * @param qm    The {@link QueryManager} to use
-     * @param cveId The CVE ID to look for
-     * @return The {@link Vulnerability} matching the CVE ID, or {@code null} when no match was found
-     */
-    private static Vulnerability getVulnerabilityByCveId(final QueryManager qm, final String cveId) {
-        final Query<Vulnerability> query = qm.getPersistenceManager().newQuery(Vulnerability.class);
-        query.setFilter("source == :source && vulnId == :cveId");
-        query.setNamedParameters(Map.of(
-                "source", Source.NVD.name(),
-                "cveId", cveId
-        ));
-        try {
-            return query.executeUnique();
-        } finally {
-            query.closeAll();
-        }
-    }
-
-    /**
-     * Update an existing, persistent {@link Vulnerability} with data as reported by the NVD.
-     * <p>
-     * It differs from {@link QueryManager#updateVulnerability(Vulnerability, boolean)} in that it keeps track of
-     * which fields are modified, and assumes the to-be-updated {@link Vulnerability} to be persistent, and enrolled
-     * in an active {@link javax.jdo.Transaction}.
-     *
-     * @param existingVuln The existing {@link Vulnerability} to update
-     * @param reportedVuln The {@link Vulnerability} as reported by the NVD
-     * @return A {@link Map} holding the differences of all updated fields
-     */
-    private static Map<String, Diff> updateVulnerability(final Vulnerability existingVuln, final Vulnerability reportedVuln) {
-        assertPersistent(existingVuln, "existingVuln must be persistent in order for changes to be effective");
-
-        final var differ = new Differ<>(existingVuln, reportedVuln);
-        differ.applyIfChanged("title", Vulnerability::getTitle, existingVuln::setTitle);
-        differ.applyIfChanged("subTitle", Vulnerability::getSubTitle, existingVuln::setSubTitle);
-        differ.applyIfChanged("description", Vulnerability::getDescription, existingVuln::setDescription);
-        differ.applyIfChanged("detail", Vulnerability::getDetail, existingVuln::setDetail);
-        differ.applyIfChanged("recommendation", Vulnerability::getRecommendation, existingVuln::setRecommendation);
-        differ.applyIfChanged("references", Vulnerability::getReferences, existingVuln::setReferences);
-        differ.applyIfChanged("credits", Vulnerability::getCredits, existingVuln::setCredits);
-        differ.applyIfChanged("created", Vulnerability::getCreated, existingVuln::setCreated);
-        differ.applyIfChanged("published", Vulnerability::getPublished, existingVuln::setPublished);
-        differ.applyIfChanged("updated", Vulnerability::getUpdated, existingVuln::setUpdated);
-        differ.applyIfNonEmptyAndChanged("cwes", Vulnerability::getCwes, existingVuln::setCwes);
-        // Calling setSeverity nulls all CVSS and OWASP RR fields. getSeverity calculates the severity on-the-fly,
-        // and will return UNASSIGNED even when no severity is set explicitly. Thus, calling setSeverity
-        // must happen before CVSS and OWASP RR fields are set, to avoid null-ing them again.
-        differ.applyIfChanged("severity", Vulnerability::getSeverity, existingVuln::setSeverity);
-        differ.applyIfChanged("cvssV2BaseScore", Vulnerability::getCvssV2BaseScore, existingVuln::setCvssV2BaseScore);
-        differ.applyIfChanged("cvssV2ImpactSubScore", Vulnerability::getCvssV2ImpactSubScore, existingVuln::setCvssV2ImpactSubScore);
-        differ.applyIfChanged("cvssV2ExploitabilitySubScore", Vulnerability::getCvssV2ExploitabilitySubScore, existingVuln::setCvssV2ExploitabilitySubScore);
-        differ.applyIfChanged("cvssV2Vector", Vulnerability::getCvssV2Vector, existingVuln::setCvssV2Vector);
-        differ.applyIfChanged("cvssV3BaseScore", Vulnerability::getCvssV3BaseScore, existingVuln::setCvssV3BaseScore);
-        differ.applyIfChanged("cvssV3ImpactSubScore", Vulnerability::getCvssV3ImpactSubScore, existingVuln::setCvssV3ImpactSubScore);
-        differ.applyIfChanged("cvssV3ExploitabilitySubScore", Vulnerability::getCvssV3ExploitabilitySubScore, existingVuln::setCvssV3ExploitabilitySubScore);
-        differ.applyIfChanged("cvssV3Vector", Vulnerability::getCvssV3Vector, existingVuln::setCvssV3Vector);
-        differ.applyIfChanged("owaspRRLikelihoodScore", Vulnerability::getOwaspRRLikelihoodScore, existingVuln::setOwaspRRLikelihoodScore);
-        differ.applyIfChanged("owaspRRTechnicalImpactScore", Vulnerability::getOwaspRRTechnicalImpactScore, existingVuln::setOwaspRRTechnicalImpactScore);
-        differ.applyIfChanged("owaspRRBusinessImpactScore", Vulnerability::getOwaspRRBusinessImpactScore, existingVuln::setOwaspRRBusinessImpactScore);
-        differ.applyIfChanged("owaspRRVector", Vulnerability::getOwaspRRVector, existingVuln::setOwaspRRVector);
-        differ.applyIfChanged("vulnerableVersions", Vulnerability::getVulnerableVersions, existingVuln::setVulnerableVersions);
-        differ.applyIfChanged("patchedVersions", Vulnerability::getPatchedVersions, existingVuln::setPatchedVersions);
-        // EPSS is an additional enrichment that no source currently provides natively. We don't want EPSS scores of CVEs to be purged.
-        differ.applyIfNonNullAndChanged("epssScore", Vulnerability::getEpssScore, existingVuln::setEpssScore);
-        differ.applyIfNonNullAndChanged("epssPercentile", Vulnerability::getEpssPercentile, existingVuln::setEpssPercentile);
-
-        return differ.getDiffs();
     }
 
     private static final class HttpClientSupplier implements HttpAsyncClientSupplier {
